requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
magic-filter==1.0.12
multidict==6.0.5
packaging==23.2
pexelsPy==1.0.1
pluggy==1.4.0
pyaes==1.6.1
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
pytest==8.0.0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
